+++ src/Recipe/Draft.php
@@ -57,6 +57,9 @@ class Draft implements StateInterface
 {
     use StateTrait;
 
+    /*
+     * To define a new required ingredient to execute the recipe, a new recipe object will be returned
+     */
     public function addIngredient(): callable
     {
         return function (IngredientInterface $ingredient): RecipeInterface {
@@ -70,6 +73,10 @@ class Draft implements StateInterface
         };
     }
 
+    /*
+     * To push a step in the recipe, a new recipe object will be returned. If the step is a callable, it will be wrapped
+     * into a Bowl object, else, the BowlInterface instance will be directly used
+     */
     public function addStep(): callable
     {
         return function (
@@ -98,6 +105,9 @@ class Draft implements StateInterface
         };
     }
 
+    /*
+     * To define an error handler about this recipe, a new recipe object will be returned
+     */
     public function setOnError(): callable
     {
         return function (callable | BowlInterface $action): RecipeInterface {
@@ -117,6 +127,11 @@ class Draft implements StateInterface
         };
     }
 
+    /*
+     * To define / add a sub recipe into this recipe recipe. It will be wrapped into a RecipeBowl instance. The repeat
+     * condition can be a callable and will be wrapped into a bowl in this case.
+     * A new recipe object will be returned
+     */
     public function addSubRecipe(): callable
     {
         return function (
@@ -146,6 +161,9 @@ class Draft implements StateInterface
         };
     }
 
+    /*
+     * To define excepted dish resulting of the recipe, a new recipe object will be returned
+     */
     public function setExceptedDish(): callable
     {
         return function (DishInterface $dish): RecipeInterface {